requirements.txt
fastapi
uvicorn
playwright
google-generativeai
requests
pandas
numpy
pdfplumber
beautifulsoup4
python-multipart
